api/requirements.txt
aiocache==0.12.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
astroid==4.0.2
async-timeout==5.0.1
bitarray==3.8.0
black==25.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.1
coverage==7.12.0
cytoolz==1.1.0
dill==0.4.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
exceptiongroup==1.3.1
fastapi==0.123.9
gevent==25.9.1
greenlet==3.3.0
h11==0.16.0
h2==4.3.0
hexbytes==1.3.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
isort==7.0.0
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
parsimonious==0.10.0
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
poly_eip712_structs==0.0.1
py_builder_signing_sdk==0.0.2
py_clob_client==0.30.0
py_order_utils==0.3.2
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pylint==4.0.4
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
redis==7.1.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rlp==4.1.0
six==1.17.0
starlette==0.50.0
tenacity==9.1.2
tomli==2.3.0
tomlkit==0.13.3
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
uvicorn==0.38.0
websocket-client==1.9.0
zope.event==6.1
zope.interface==8.1.1
